#5328 simplifies changes to just replace the MemorySession adapter based on the internal phalcon version
 - changes private accessors for memory session
 - removes all fluff for memory-session override
